+++ stroom-query-api/src/main/java/stroom/query/api/v2/DateTimeFormat.java
@@ -17,6 +17,8 @@
 package stroom.query.api.v2;
 
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -24,29 +26,51 @@ import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlType;
 import java.io.Serializable;
 
+/**
+ * Class for describing the format to use for formatting a date time value
+ */
 @JsonPropertyOrder({"pattern", "timeZone"})
 @XmlType(name = "DateTimeFormat", propOrder = {"pattern", "timeZone"})
 @XmlAccessorType(XmlAccessType.FIELD)
+@ApiModel(description = "The string formatting to apply to a date value")
 public final class DateTimeFormat implements Serializable {
     private static final long serialVersionUID = 9145624653060319801L;
 
     @XmlElement
+    @ApiModelProperty(
+            value = "A date time formatting pattern string conforming to the specification of " +
+                    "java.time.format.DateTimeFormatter",
+            required = true)
     private String pattern;
+
     @XmlElement
+    @ApiModelProperty(
+            required = true)
     private TimeZone timeZone;
 
     private DateTimeFormat() {
     }
 
+    /**
+     * @param pattern A date time formatting pattern string conforming to the specification of
+     * {@link java.time.format.DateTimeFormatter}
+     * @param timeZone The time zone to use when formatting the date time value
+     */
     public DateTimeFormat(final String pattern, final TimeZone timeZone) {
         this.pattern = pattern;
         this.timeZone = timeZone;
     }
 
+    /**
+     * @return The format pattern string, conforming to {@link java.time.format.DateTimeFormatter}
+     */
     public String getPattern() {
         return pattern;
     }
 
+    /**
+     * @return The the {@link TimeZone timeZone} to use when formatting the date
+     */
     public TimeZone getTimeZone() {
         return timeZone;
     }