+++ stroom-query-api/src/main/java/stroom/query/api/v2/Query.java
@@ -17,6 +17,8 @@
 package stroom.query.api.v2;
 
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -27,19 +29,36 @@ import javax.xml.bind.annotation.XmlType;
 import java.io.Serializable;
 import java.util.List;
 
+/**
+ * {@value #CLASS_DESC}
+ */
 @JsonPropertyOrder({"dataSource", "expression", "params"})
 @XmlType(name = "Query", propOrder = {"dataSource", "expression", "params"})
 @XmlRootElement(name = "query")
 @XmlAccessorType(XmlAccessType.FIELD)
+@ApiModel(description = Query.CLASS_DESC)
 public final class Query implements Serializable {
+
     private static final long serialVersionUID = 9055582579670841979L;
 
+    public static final String CLASS_DESC = "The query terms for the search";
+
     @XmlElement
+    @ApiModelProperty(
+            required = true)
     private DocRef dataSource;
+
     @XmlElement
+    @ApiModelProperty(
+            value = "The root logical operator in the query expression tree",
+            required = true)
     private ExpressionOperator expression;
+
     @XmlElementWrapper(name = "params")
     @XmlElement(name = "param")
+    @ApiModelProperty(
+            value = "A list of key/value pairs that provide additional information about the query",
+            required = false)
     private List<Param> params;
 
     private Query() {