+++ stroom-query-api/src/main/java/stroom/query/api/v2/SearchRequest.java
@@ -17,6 +17,8 @@
 package stroom.query.api.v2;
 
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -27,29 +29,71 @@ import javax.xml.bind.annotation.XmlType;
 import java.io.Serializable;
 import java.util.List;
 
+/**
+ * A class for describing a search request including the query to run and definition(s) of how the results
+ * should be returned
+ */
 @JsonPropertyOrder({"key", "query", "resultRequests", "dateTimeLocale", "incremental"})
 @XmlRootElement(name = "searchRequest")
 @XmlType(name = "SearchRequest", propOrder = {"key", "query", "resultRequests", "dateTimeLocale", "incremental"})
 @XmlAccessorType(XmlAccessType.FIELD)
+@ApiModel(description = "A request for new search or a follow up request for more data for an existing iterative search")
 public final class SearchRequest implements Serializable {
     private static final long serialVersionUID = -6668626615097471925L;
 
     @XmlElement
+    @ApiModelProperty(required = true)
     private QueryKey key;
+
     @XmlElement
+    @ApiModelProperty(required = true)
     private Query query;
+
     @XmlElementWrapper(name = "resultRequests")
     @XmlElement(name = "resultRequest")
+    @ApiModelProperty(required = true)
     private List<ResultRequest> resultRequests;
+
+
     @XmlElement
+    @ApiModelProperty(
+            value = "The locale to use when formatting date values in the search results. The " +
+                    "value is the string form of a java.time.ZoneId",
+            required = true)
     private String dateTimeLocale;
+
     @XmlElement
+    @ApiModelProperty(
+            value = "If true the response will contain all results found so far. Future requests " +
+            "for the same query key may return more results. Intended for use on longer running searches to allow " +
+            "partial result sets to be returned as soon as they are available rather than waiting for the full " +
+            "result set.",
+            required = true)
     private Boolean incremental;
 
     private SearchRequest() {
     }
 
-    public SearchRequest(final QueryKey key, final Query query, final List<ResultRequest> resultRequests, final String dateTimeLocale, final Boolean incremental) {
+    /**
+     * @param key            A unique key to identify the instance of the search by. This key is used to identify multiple
+     *                       requests for the same search when running in incremental mode.
+     * @param query          The query terms for the search
+     * @param resultRequests A list of {@link ResultRequest resultRequest} definitions. If null or the list is empty
+     *                       no results will be returned. Allows the caller to request that the results of the query
+     *                       are returned in multiple forms, e.g. using a number of different
+     *                       filtering/aggregation/sorting approaches.
+     * @param dateTimeLocale The locale to use when formatting date values in the search results. The value is the
+     *                       string form of a {@link java.time.ZoneId zoneId}
+     * @param incremental    If true the response will contain all results found so far. Future requests for the same
+     *                       query key may return more results. Intended for use on longer running searches to allow
+     *                       partial result sets to be returned as soon as they are available rather than waiting for the
+     *                       full result set.
+     */
+    public SearchRequest(final QueryKey key,
+                         final Query query,
+                         final List<ResultRequest> resultRequests,
+                         final String dateTimeLocale,
+                         final Boolean incremental) {
         this.key = key;
         this.query = query;
         this.resultRequests = resultRequests;
@@ -57,26 +101,46 @@ public final class SearchRequest implements Serializable {
         this.incremental = incremental;
     }
 
+    /**
+     * @return The unique {@link QueryKey queryKey} for the search request
+     */
     public QueryKey getKey() {
         return key;
     }
 
+    /**
+     * @return The {@link Query query} object containing the search terms
+     */
     public Query getQuery() {
         return query;
     }
 
+    /**
+     * @return The list of {@link ResultRequest resultRequest} objects
+     */
     public List<ResultRequest> getResultRequests() {
         return resultRequests;
     }
 
+    /**
+     * @return The locale ID, see {@link java.time.ZoneId}, for the date values uses in the search response.
+     */
     public String getDateTimeLocale() {
         return dateTimeLocale;
     }
 
+    /**
+     * @return Whether the search should return immediately with the results found so far or wait for the search
+     * to finish
+     */
     public Boolean getIncremental() {
         return incremental;
     }
 
+    /**
+     * @return Whether the search should return immediately with the results found so far or wait for the search
+     * to finish
+     */
     public boolean incremental() {
         return incremental != null && incremental;
     }