+++ stroom-query-api/src/main/java/stroom/query/api/v2/Sort.java
@@ -17,6 +17,8 @@
 package stroom.query.api.v2;
 
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
 import stroom.util.shared.HasDisplayValue;
 
 import javax.xml.bind.annotation.XmlAccessType;
@@ -28,12 +30,23 @@ import java.io.Serializable;
 @JsonPropertyOrder({"order", "direction"})
 @XmlType(name = "Sort", propOrder = {"order", "direction"})
 @XmlAccessorType(XmlAccessType.FIELD)
+@ApiModel(description = "Describes the sorting applied to a field")
 public final class Sort implements Serializable {
     private static final long serialVersionUID = 4530846367973824427L;
 
     @XmlElement
+    @ApiModelProperty(
+            value = "Where multiple fields are sorted this value describes the sort order, with 0 being the first " +
+                    "field to sort on",
+            example = "0",
+            required = true)
     private Integer order;
+
     @XmlElement
+    @ApiModelProperty(
+            value = "The direction to sort in, ASCENDING or DESCENDING",
+            example = "ASCENDING",
+            required = true)
     private SortDirection direction;
 
     private Sort() {